fastapi/requirements.txt
# FastAPI and its dependencies
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0

# Database
sqlalchemy>=1.4.0,<1.5.0

# CORS middleware (included with FastAPI/Starlette)
starlette>=0.14.2,<0.15.0

# Authentication dependencies
python-jose[cryptography]>=3.3.0,<4.0.0  # JWT token handling
passlib[bcrypt]>=1.7.4,<2.0.0              # Password hashing
python-multipart>=0.0.5,<0.1.0             # Form data parsing for OAuth2

# Development dependencies
python-dotenv>=0.19.0,<0.20.0  # For environment variable management
